IOS/USB: Fix TransferCommand length type
The total buffer size for isochronous transfers should be a u32,
not a u16. It is easy to hit the bug with devices such as cameras,
which require larger buffers.

This fixes Your Shape.

This also fixes the length type for bulk and interrupt transfers,
which should be u32 as that's what IOS supports. I'm not sure why
I made them u16... probably because OH0 uses u16 for most lengths...

+++ Source/Core/Core/IOS/USB/Common.h
@@ -127,14 +127,14 @@ struct CtrlMessage : TransferCommand
 
 struct BulkMessage : TransferCommand
 {
-  u16 length = 0;
+  u32 length = 0;
   u8 endpoint = 0;
   using TransferCommand::TransferCommand;
 };
 
 struct IntrMessage : TransferCommand
 {
-  u16 length = 0;
+  u32 length = 0;
   u8 endpoint = 0;
   using TransferCommand::TransferCommand;
 };
@@ -143,7 +143,7 @@ struct IsoMessage : TransferCommand
 {
   u32 packet_sizes_addr = 0;
   std::vector<u16> packet_sizes;
-  u16 length = 0;
+  u32 length = 0;
   u8 num_packets = 0;
   u8 endpoint = 0;
   using TransferCommand::TransferCommand;
